requirements.txt
python-dotenv
flask-sqlalchemy
flask-migrate
apscheduler
selenium